requirements.txt
asgiref==3.5.2
Django==4.1.1
djangorestframework==3.13.1
psycopg2-binary==2.9.3
PyJWT==2.4.0
pytz==2022.2.1
sqlparse==0.4.2
